requirements.txt
openai==1.30.3
playwright==1.34.0
python-dotenv==1.0.1
setuptools==65.6.3